requirements.txt
alembic==1.17.0
asyncpg==0.30.0
fastapi==0.116.1
pyjwt==2.10.1
pydantic==2.11.7
sqlalchemy==2.0.42